--- src/main/java/app/persistence/Team05.java
@@ -1,4 +1,0 @@
-package app.persistence;
-
-public class Team05 {
-}
